+++ pkg/analyzer/lib/dart/element/element.dart
@@ -2289,6 +2289,7 @@ abstract class _TmpSharedClassElement {
   bool get hasNonFinalField;
 
   /// Return `true` if this class declares a static member.
+  @Deprecated('Not useful for clients')
   bool get hasStaticMember;
 
   /// Return `true` if this class is abstract. A class is abstract if it has an
@@ -2327,7 +2328,7 @@ abstract class _TmpSharedClassElement {
   /// declares a constructor. It is a compile-time error if a mixin is derived
   /// from a class whose superclass is not Object.
   /// </blockquote>
-  /// TODO(scheglov) Deprecate and remove it.
+  @Deprecated('Not useful for clients')
   bool get isValidMixin;
 
   /// Return a list containing all of the superclass constraints defined for